reorganizing workout screen
